+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.5.0] - 2026-03-29
+
 ### Added
 
 - hnefatafl-client: maximum and minimum rating for archived games.
@@ -728,7 +730,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://github.com/dcampbell24/hnefatafl/compare/v5.4.0-1...main
+[unreleased]: https://github.com/dcampbell24/hnefatafl/compare/v5.5.0-1...main
+[5.5.0]: https://github.com/dcampbell24/hnefatafl/compare/v5.4.0-1...v5.0.0-1
 [5.4.0]: https://github.com/dcampbell24/hnefatafl/compare/v5.3.1-1...v5.4.0-1
 [5.3.1]: https://github.com/dcampbell24/hnefatafl/compare/v5.3.0-2...v5.3.1-1
 [5.3.0]: https://github.com/dcampbell24/hnefatafl/compare/v5.2.0-1...v5.3.0-2